backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
langgraph==0.2.66
langsmith==0.1.147
pydantic==2.9.2
python-dotenv==1.0.1
